+++ packages/@headlessui-react/jest.config.cjs
@@ -1,2 +1,5 @@
 let create = require('../../jest/create-jest-config.cjs')
-module.exports = create(__dirname, { displayName: 'React' })
+module.exports = create(__dirname, {
+  displayName: 'React',
+  setupFilesAfterEnv: ['./jest.setup.js'],
+})