+++ packages/@headlessui-react/src/components/disclosure/disclosure.test.tsx
@@ -12,7 +12,7 @@ import {
   assertActiveElement,
   getByText,
 } from '../../test-utils/accessibility-assertions'
-import { click, press, Keys, MouseButton } from '../../test-utils/interactions'
+import { click, press, focus, Keys, MouseButton } from '../../test-utils/interactions'
 import { Transition } from '../transitions/transition'
 
 jest.mock('../../hooks/use-id')
@@ -133,7 +133,7 @@ describe('Rendering', () => {
         )
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -174,7 +174,7 @@ describe('Rendering', () => {
         )
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -218,7 +218,7 @@ describe('Rendering', () => {
         render(<Example />)
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -437,7 +437,7 @@ describe('Rendering', () => {
         )
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -474,7 +474,7 @@ describe('Rendering', () => {
         )
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -514,7 +514,7 @@ describe('Rendering', () => {
         render(<Example />)
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Ensure the button is focused
         assertActiveElement(getDisclosureButton())
@@ -579,14 +579,15 @@ describe('Composition', () => {
 
       // Wait for all transitions to finish
       await nextFrame()
+      await nextFrame()
 
       // Verify that we tracked the `mounts` and `unmounts` in the correct order
       expect(orderFn.mock.calls).toEqual([
         ['Mounting - Disclosure'],
         ['Mounting - Transition'],
         ['Mounting - Transition.Child'],
-        ['Unmounting - Transition.Child'],
         ['Unmounting - Transition'],
+        ['Unmounting - Transition.Child'],
       ])
     })
   )
@@ -611,7 +612,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Open disclosure
         await press(Keys.Enter)
@@ -646,7 +647,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Try to open the disclosure
         await press(Keys.Enter)
@@ -677,7 +678,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Open disclosure
         await press(Keys.Enter)
@@ -717,7 +718,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Open disclosure
         await press(Keys.Space)
@@ -748,7 +749,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Try to open the disclosure
         await press(Keys.Space)
@@ -779,7 +780,7 @@ describe('Keyboard interactions', () => {
         assertDisclosurePanel({ state: DisclosureState.InvisibleUnmounted })
 
         // Focus the button
-        getDisclosureButton()?.focus()
+        await focus(getDisclosureButton())
 
         // Open disclosure
         await press(Keys.Space)