+++ packages/@headlessui-react/src/hooks/use-latest-value.ts
@@ -1,9 +1,10 @@
-import { useRef, useEffect } from 'react'
+import { useRef } from 'react'
+import { useIsoMorphicEffect } from './use-iso-morphic-effect'
 
 export function useLatestValue<T>(value: T) {
   let cache = useRef(value)
 
-  useEffect(() => {
+  useIsoMorphicEffect(() => {
     cache.current = value
   }, [value])