+++ packages/@headlessui-react/src/test-utils/interactions.ts
@@ -4,11 +4,13 @@ import { disposables } from '../utils/disposables'
 let d = disposables()
 
 function nextFrame(cb: Function): void {
-  setImmediate(() =>
+  setImmediate(() => {
     setImmediate(() => {
-      cb()
+      setImmediate(() => {
+        cb()
+      })
     })
-  )
+  })
 }
 
 export let Keys: Record<string, Partial<KeyboardEvent>> = {
@@ -282,7 +284,11 @@ export async function focus(element: Document | Element | Window | Node | null) 
   try {
     if (element === null) return expect(element).not.toBe(null)
 
-    fireEvent.focus(element)
+    if (element instanceof HTMLElement) {
+      element.focus()
+    } else {
+      fireEvent.focus(element)
+    }
 
     await new Promise(nextFrame)
   } catch (err) {