+++ packages/playground-react/next.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  reactStrictMode: true,
   devIndicators: {
     autoPrerender: false,
   },